src/requirements.txt
pypdf
langchain
langchain-community
langchain-openai
openai
tiktoken
faiss-cpu
flask